fix collection binding file contents
